+++ libs/shared/ofrep-core/src/lib/api/ofrep-api.spec.ts
@@ -20,12 +20,21 @@ import {
 describe('OFREPApi', () => {
   let api: OFREPApi;
 
-  beforeAll(() => server.listen());
+  beforeAll(() => {
+    server.listen();
+  });
   beforeEach(() => {
+    jest.useFakeTimers();
     api = new OFREPApi('https://localhost:8080');
   });
-  afterEach(() => server.resetHandlers());
-  afterAll(() => server.close());
+  afterEach(() => {
+    jest.runOnlyPendingTimers();
+    jest.useRealTimers();
+    server.resetHandlers();
+  });
+  afterAll(() => {
+    server.close();
+  });
 
   describe('postEvaluateFlags should', () => {
     it('throw OFREPApiFetchError on network error', async () => {
@@ -56,6 +65,51 @@ describe('OFREPApi', () => {
       await expect(() => api.postEvaluateFlags('my-flag', { context: { errors: { 429: true } } })).rejects.toThrow(
         OFREPApiTooManyRequestsError,
       );
+    });
+
+    it('parse numeric Retry-After header correctly on 429 response', async () => {
+      jest.setSystemTime(new Date('2018-01-27'));
+
+      try {
+        await api.postEvaluateFlags('my-flag', { context: { errors: { 429: true } } });
+      } catch (error) {
+        if (!(error instanceof OFREPApiTooManyRequestsError)) {
+          throw new Error('Expected OFREPApiTooManyRequestsError');
+        }
+
+        expect(error.retryAfterSeconds).toEqual(2000);
+        expect(error.retryAfterDate).toEqual(new Date('2018-01-27T00:33:20.000Z'));
+      }
+    });
+
+    it('parse date Retry-After header correctly on 429 response', async () => {
+      jest.setSystemTime(new Date('2018-01-27'));
+
+      try {
+        await api.postEvaluateFlags('my-flag', { context: { errors: { 429: 'Sat, 27 Jan 2018 07:28:00 GMT' } } });
+      } catch (error) {
+        if (!(error instanceof OFREPApiTooManyRequestsError)) {
+          throw new Error('Expected OFREPApiTooManyRequestsError');
+        }
+
+        expect(error.retryAfterSeconds).toEqual(null);
+        expect(error.retryAfterDate).toEqual(new Date('2018-01-27T07:28:00.000Z'));
+      }
+    });
+
+    it('ignore Retry-After header if it is not valid on 429 response', async () => {
+      jest.setSystemTime(new Date('2018-01-27'));
+
+      try {
+        await api.postEvaluateFlags('my-flag', { context: { errors: { 429: 'abcdefg' } } });
+      } catch (error) {
+        if (!(error instanceof OFREPApiTooManyRequestsError)) {
+          throw new Error('Expected OFREPApiTooManyRequestsError');
+        }
+
+        expect(error.retryAfterSeconds).toEqual(null);
+        expect(error.retryAfterDate).toEqual(null);
+      }
     });
 
     it('send empty request body if context is not given', async () => {