+++ libs/shared/ofrep-core/src/lib/provider/index.ts
@@ -1,0 +1,1 @@
+export * from './ofrep-provider-options';